edit open order catching

+++ src/commands/orders/deleteOrder.ts
@@ -8,6 +8,7 @@
 import * as BackendJS from "backendjs";
 import * as CoreJS from "corejs";
 import { Args as GlobalArgs, Context, Options } from "../../core";
+import { OrderState } from "../../enums";
 
 interface Args extends GlobalArgs {
     readonly account: number;
@@ -29,6 +30,9 @@ export class DeleteOrder extends BackendJS.Module.Command<Context, Args, Options
 
         if (order.account != args.account)
             return new CoreJS.ErrorResponse(CoreJS.ResponseCode.Forbidden, '#_permission_denied');
+
+        if (order.state != OrderState.Open)
+            return new CoreJS.ErrorResponse(CoreJS.ResponseCode.Forbidden, '#_order_not_open');
 
         const result = await this.context.orderRepository.deleteOrder(args.order);
